pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "photo-restore"
version = "0.1.0"
description = "AI-powered photo restoration CLI tool"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Photo Restore Team", email = "team@photo-restore.com"},
]
keywords = ["photo", "restoration", "image", "enhancement", "AI", "CLI"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "torch>=1.9.0",
    "opencv-python>=4.5.0",
    "Pillow>=8.0.0",
    "numpy>=1.21.0",
    "click>=8.0.0",
    "tqdm>=4.62.0",
    "PyYAML>=5.4.0",
    "realesrgan>=0.3.0",
    "gfpgan>=1.3.0",
    "basicsr>=1.4.0",
    "facexlib>=0.2.5",
    "requests>=2.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.0.0",
    "black>=21.0.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
]

[project.scripts]
photo-restore = "photo_restore.cli:main"

[project.urls]
Homepage = "https://github.com/username/photo-restore"
Documentation = "https://github.com/username/photo-restore#readme"
Repository = "https://github.com/username/photo-restore"
"Bug Tracker" = "https://github.com/username/photo-restore/issues"

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=photo_restore --cov-report=term-missing --cov-report=html"